==> code/espejo/past/01-data-cleaning.ipynb <==
**If you are going to re-run this code, please do not overwrite our csvs... Also, paths have been updated, but not save paths.**

# Series data
I have taken the series data from genbank. I will clean it.

library(tidyverse)

clean_af <- read_csv("../../data/genbank/original/excel-to-csv/series.csv")
head(clean_af)

I want samples to be the rows, so I will transpose.

test <- t(clean_af)
test <- data.frame(test)

names(test) <- unlist(test[1,])
test <- test[-1,]
head(test)

dim(test)

# write.csv(test, "../../data/clean/series_clean.csv")

# Metadata
I am going to clean the metadata.

metadata <- read_csv("../../data/genbank/original/excel-to-csv/sample_metadata.csv")

metadata <- rbind(names(metadata), metadata)
class(metadata)

names(metadata) <- c("attribute", metadata[2,2:ncol(metadata)])

metadata <- t(metadata[-2,])
metadata <- data.frame(metadata)
names(metadata) <- unlist(metadata[1,])
names(metadata) <- gsub("!", "", names(metadata))
metadata <- metadata[-1,]

names(metadata)[which(grepl("Sample_characteristics_ch1", names(metadata)))] <- paste0("Sample_characteristics_ch1_", 1:3)

head(metadata)

names(metadata)

desired_metadata <- metadata %>% select(Sample_characteristics_ch1_1, Sample_characteristics_ch1_2, Sample_characteristics_ch1_3)
# desired_metadata

desired_metadata <- desired_metadata %>% rename(tissue=Sample_characteristics_ch1_1,
                            endometriosis=Sample_characteristics_ch1_2,
                            severity=Sample_characteristics_ch1_3)

desired_metadata$tissue <- gsub("tissue: ", "", desired_metadata$tissue)

desired_metadata$endometriosis <- gsub("endometriosis/no endometriosis: ", "", desired_metadata$endometriosis)

desired_metadata$severity <- gsub("endometriosis severity: ", "", desired_metadata$severity)
desired_metadata$severity <- gsub("presence or absence of uterine/pelvic pathology: ", "", desired_metadata$severity)

head(desired_metadata)

# write.csv(desired_metadata, "../../data/clean/metadata.csv")

# How big is this dataset?

series <- test
dim(test)

As stated in the study, **n=148**.

table(desired_metadata_2[,2:3])

# Merging

I will merge the dataset, but you may just simply want to work on series alone then add in metadata later.

merged_df <- merge(desired_metadata, test, by=0, all=TRUE)

names(merged_df)[1] <- "sample"

head(merged_df)

We can now play around with this use `dplyr` or the like.

# write.csv(merged_df, "../../data/clean/endometriosis.csv")

# Subset
I want to play with a subset of the data only. My computer might burst into flames with such a big dataset.

ix <- sample(x=1:ncol(merged_df), size=100, replace=FALSE)

48311 48038 46259 45607 30859 36212 27412 49211 12498 37608 17275 8977 51860 6773 4279 41073 52002 43853 30942 12710 14552 4363 2390 18347 36729 4004 25003 40453 85 21466 15258 50398 8650 40258 26186 25705 32507 49416 34377 47942 17768 32303 15772 12870 47486 2577 7727 24999 39689 7776 47238 12204 51666 42631 45989 39079 36200 11922 35675 37843 32106 41530 10603 48845 46623 46727 32260 13626 36846 41145 26506 23904 29141 44410 10340 11948 15524 24035 18054 20767 42290 45802 7007 17957 27856 31556 40242 44022 8213 12002 2864 24337 19793 37263 41779 29699 6940 991 23800 18913

# ix

small_subset <-merged_df[,c(1:4,ix)]

# write.csv(small_subset, "../../data/clean/small_subset.csv")
