==> code/espejo/past/01-dataset-summaries.ipynb <==
# Genbank Dataset

import pandas as pd
import numpy as np
import os

series = pd.read_csv("../../data/genbank/clean-csv/series_clean.csv")

series.shape

# Figshare Dataset

series_2a = pd.read_csv("../../data/figshare/clean-csv/minimal-mild.csv")

series_2b = pd.read_csv("../../data/figshare/clean-csv/moderate-severe.csv")

series_2c = pd.read_csv("../../data/figshare/clean-csv/normal.csv")

series_2 = series_2a.append([series_2b, series_2c])

series_2.shape

# Subject Names

same_subjects = np.intersect1d(series[series.columns[0]], series_2[series_2.columns[0]])

len(same_subjects)

All of the subjects within the Figshare dataset are within the Genbank dataset.

# Compare Information for One Subject

series[series.columns[0]][0]

I am just choosing the first subject. I want to see if the Figshare data is entirely included in the genbank data.

series_2.loc[series_2[series_2.columns[0]] == "GSM1256653"]

pd.DataFrame(series.iloc[0]).T

pd.DataFrame(series.iloc[0])

# Check Sample

pd.DataFrame(series.iloc[0]).T["121_at"]

series_2.loc[series_2[series_2.columns[0]] == "GSM1256653"]["PAX8"]
