==> code/espejo/past/01-probe-functions.ipynb <==
http://xavierlab2.mgh.harvard.edu/EnrichmentProfiler/cancer/Expression/AFFX-r2-Ec-bioB-3_at.html

https://blog.rstudio.com/2015/04/21/xml2/

https://rdrr.io/cran/xml2/man/xml_name.html

library(xml2)

# web_text <- scan("http://xavierlab2.mgh.harvard.edu/EnrichmentProfiler/cancer/Expression/AFFX-r2-Ec-bioB-3_at.html", what="character")

site <- "http://xavierlab2.mgh.harvard.edu/EnrichmentProfiler/cancer/Expression/AFFX-r2-Ec-bioB-3_at.html"
site_read <- read_html(site)
site_read

xml_children(site_read)[2]

strsplit(as.character(xml_children(site_read)[2]), "<!-- enter notes -->")[[1]][2]

xml_children(xml_children(site_read)[2])[7]

as.character(xml_children(xml_children(site_read)[2])[7])

xml_attr(site_read, "id")

xml_chidren(xml_children(site_read)[2])

xml_children(site_read)[2]

line_interest <- strsplit(as.character(xml_children(xml_children(site_read)[2])[7]), 'div id=\\"notes')[[1]][2]
line_interest

ix <- lapply(strsplit(line_interest, ">")[[1]], function(x) grepl("probe", x))

strsplit(line_interest, ">")[[1]][unlist(ix)]

# Function to Ask the Site

probe_to_gene <- function(probe_name) {
    # * USING THIS LAB INFO
    site <- "http://xavierlab2.mgh.harvard.edu/EnrichmentProfiler/cancer/Expression/"
    site <- paste0(site, probe_name, ".html")
    site_read <- read_html(site)
    
    if (length(site_read)>0) {
        line_interest <- strsplit(as.character(xml_children(xml_children(site_read)[2])[7]), 'div id=\\"notes')[[1]][2]
        
        ix <- lapply(strsplit(line_interest, ">")[[1]], function(x) grepl("probe", x))
        site <- strsplit(line_interest, ">")[[1]][unlist(ix)][1]
                     
        if(!grepl("control", site)) {
              site <- strsplit(site, '\\"')[[1]][2]
            site_read <- read_html(site)
            gene <- strsplit(as.character(xml_children(site_read)[1]), "<title>")
            strsplit(strsplit(gene[[1]][2], "</title>")[[1]][1], " ")[[1]][1]
        } else {
            "control"
        }
        
                    
                     
    } else {
        return(0)
    }
}

probe_to_gene("1053_at")

probe_to_gene("117_at")

probe_to_gene("121_at")

probe_to_gene("AFFX-TrpnX-3_at")
