==> code/espejo/past/02-cut-up-data.ipynb <==
# Subsets for Control Removal

import pandas as pd
import numpy as np

endo = pd.read_csv("../../data/genbank/clean-csv/endometriosis.csv")

endo = endo.drop(endo.columns[0], axis=1)

endo.shape

endo.shape[1]-4

ranges = [[np.min(x), np.max(x)] for x in [range(y*(endo.shape[1]-4)/10,(y+1)*(endo.shape[1]-4)/10) for y in range(0,10)]]
ranges

# ranges[len(ranges)-1][1] = ranges[len(ranges)-1][1]+1

ranges = pd.DataFrame(ranges, columns=["start", "end"])
ranges

All of these indices are off by "a count of 4". These four columns will be included in all subsets.

meta = endo[endo.columns[0:4]]
meta.head()

this_range = range(list(ranges.iloc[0])[0], list(ranges.iloc[0])[1])
this_cut   = endo[endo.columns[[x+4 for x in this_range]]]
this_cut.head()

this_subset = meta.join(this_cut, how="outer")
this_subset.head()

this_file = "../../data/genbank/cut_1_test.csv"
this_subset.to_csv(this_file, encoding="utf-8")

def cut_csv(ranges, out_name):
    this_range = range(ranges[0], ranges[1])
    this_cut   = endo[endo.columns[[x+4 for x in this_range]]]
    this_subset = meta.join(this_cut, how="outer")
    this_file = "../../data/genbank/subsets-for-control-removal/" + out_name + ".csv"
    this_subset.to_csv(this_file, encoding="utf-8")  

out_names = ["cut_"+str(x) for x in range(1,11)]
out_names

for index, row in ranges.iterrows():
    print(index)
    cut_csv(list(row), out_names[index])
